requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Flask-JWT-Extended>=4.6.0
Flask-CORS>=4.0.0
PyMySQL>=1.1.0
cryptography>=42.0.0
python-dotenv>=1.0.0
marshmallow>=3.21.0
Werkzeug>=3.0.0
pypdf>=4.0.0
python-docx>=1.1.0
openpyxl>=3.1.0
requests>=2.31.0
